requirements.txt
celery==5.2.7
Django==3.2.7
django-cacheops==6.1
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-health-check==3.16.5
django-ratelimit==3.0.1
django-rest-passwordreset==1.2.1
django-ses==3.1.0
django-simple-history==3.1.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flower==1.1.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
redis==4.3.4
stripe==3.5.0